Clearing up "creating a node" ambiguity
You mention node later on (in line 103) but don't introduce it anywhere. You then revisit adding a nodes in the "Life Inside a Cluster" section but users are still ambiguous on nodes. Adding "node" to line 67 will prepare users to refer back to this point when they want to create a new node and introduce the concepts of nodes to them.